Add option to disable logging to file on Linux
and use that option for the systemd service. We're going to rely on
systemd's journal logs from now on instead of creating our own log
files. This has the advantage that we don't need to care about
rotation and deletion of old files as this will all be handled by
systemd-journald.