Change balance device driver to be more modular
There is a new generic `BalanceInterface` class now which contains the
minimal interface a balance must have to be compatibale with the
`BalanceService` Feature. Additionally, there is a
`SerialBalanceInterface` that adds methods to ease communicating with a
balance over a serial connection. The specific device driver
implementation is now in 'sartorius_balance.py' and is still called
`SartoriusBalance`. The `BalanceService` now only uses the
`BalanceInterface` which allows to easily change the specific driver
implementation.